Split CarbonBased and SparkBased tests
